fixed issue in workshop mod update checked: "Could not convert to integer"

+++ src/Bot/Util/SteamWebAPI.cs
@@ -27,7 +27,7 @@ public static class SteamWebAPI
             public int    ConsumerAppId;
             public string FileName;
             [JsonProperty("file_size")]
-            public int    FileSize;
+            public long   FileSize;
             [JsonProperty("file_url")]
             public string FileURL;
             [JsonProperty("hcontent_file")]
